change float back to double

+++ stencil.c
@@ -6,9 +6,9 @@
 #define OUTPUT_FILE "stencil.pgm"
 
 void stencil(const int nx, const int ny, const int width, const int height,
-             float* image, float* tmp_image);
+             double* image, double* tmp_image);
 void init_image(const int nx, const int ny, const int width, const int height,
-                float* image, float* tmp_image);
+                double* image, double* tmp_image);
 void output_image(const char* file_name, const int nx, const int ny,
                   const int width, const int height, float* image);
 double wtime(void);
@@ -32,8 +32,9 @@ int main(int argc, char* argv[])
   int height = ny + 2;
 
   // Allocate the image
-  float* image = malloc(sizeof(float) * width * height);
-  float* tmp_image = malloc(sizeof(float) * width * height);
+  // at
+  double* image = malloc(sizeof(double) * width * height);
+  double* tmp_image = malloc(sizeof(double) * width * height);
 
   // Set the input image
   init_image(nx, ny, width, height, image, tmp_image);
@@ -57,7 +58,7 @@ int main(int argc, char* argv[])
 }
 
 void stencil(const int nx, const int ny, const int width, const int height,
-             float* image, float* tmp_image)
+             double* image, double* tmp_image)
 {
   for (int j = 1; j < ny + 1; ++j) {
     for (int i = 1; i < nx + 1; ++i) {
@@ -72,7 +73,7 @@ void stencil(const int nx, const int ny, const int width, const int height,
 
 // Create the input image
 void init_image(const int nx, const int ny, const int width, const int height,
-                float* image, float* tmp_image)
+                double* image, double* tmp_image)
 {
   // Zero everything
   for (int j = 0; j < ny + 2; ++j) {
@@ -101,7 +102,7 @@ void init_image(const int nx, const int ny, const int width, const int height,
 
 // Routine to output the image in Netpbm grayscale binary image format
 void output_image(const char* file_name, const int nx, const int ny,
-                  const int width, const int height, float* image)
+                  const int width, const int height, double* image)
 {
   // Open output file
   FILE* fp = fopen(file_name, "w");
@@ -142,4 +143,3 @@ double wtime(void)
   return tv.tv_sec + tv.tv_usec * 1e-6;
 }
 
-//test new branch
